requirements.txt
bs4
requests
discord-webhook
python-dotenv